🚧 P4J3DQ task: split context upgrade coverage
