+++ packages/agentplane/src/cli/run-cli.core.upgrade.test.ts
@@ -48,7 +48,6 @@ import {
 } from "@agentplane/testkit";
 import { resolveUpdateCheckCachePath } from "./update-check.js";
 import * as prompts from "./prompts.js";
-import { withContextPolicyGatewayText } from "../shared/policy-gateway.js";
 
 function normalizeSlashes(value: string): string {
   return value.replaceAll("\\", "/");
@@ -156,89 +155,6 @@ describe("runCli", () => {
     expect(policyText).toContain("Rendered policy body.");
     expect(policyText).not.toContain("ap:fragment");
   });
-
-  it(
-    "upgrade removes context policy references for workspaces without context layer",
-    { timeout: WORKFLOW_RUNTIME_ARTIFACTS_TIMEOUT_MS },
-    async () => {
-      const root = await mkGitRepoRoot();
-      await configureGitUser(root);
-
-      let io = captureStdIO();
-      try {
-        expect(await runCli(["init", "--yes", "--root", root])).toBe(0);
-      } finally {
-        io.restore();
-      }
-
-      const gatewayPath = path.join(root, "AGENTS.md");
-      const gatewayText = await readFile(gatewayPath, "utf8");
-      await writeFile(gatewayPath, withContextPolicyGatewayText(gatewayText), "utf8");
-      await rm(path.join(root, ".agentplane", "policy", "context.must.md"), { force: true });
-      await commitAll(root, "fixture: stale context gateway without context layer");
-
-      io = captureStdIO();
-      try {
-        expect(await runCli(["upgrade", "--yes", "--root", root])).toBe(0);
-      } finally {
-        io.restore();
-      }
-
-      const upgradedGateway = await readFile(gatewayPath, "utf8");
-      expect(upgradedGateway).not.toContain("@.agentplane/policy/context.must.md");
-      await expect(
-        pathExists(path.join(root, ".agentplane", "policy", "context.must.md")),
-      ).resolves.toBe(false);
-
-      const execFileAsync = promisify(execFile);
-      const { stdout } = await execFileAsync("node", [".agentplane/policy/check-routing.mjs"], {
-        cwd: root,
-        env: cleanGitEnv(),
-      });
-      expect(String(stdout ?? "")).toContain("policy routing OK");
-    },
-  );
-
-  it(
-    "upgrade installs context policy only for initialized context workspaces",
-    { timeout: WORKFLOW_RUNTIME_ARTIFACTS_TIMEOUT_MS },
-    async () => {
-      const root = await mkGitRepoRoot();
-      await configureGitUser(root);
-
-      let io = captureStdIO();
-      try {
-        expect(await runCli(["init", "--yes", "--root", root])).toBe(0);
-        expect(await runCli(["context", "init", "--root", root])).toBe(0);
-      } finally {
-        io.restore();
-      }
-
-      const contextPolicyPath = path.join(root, ".agentplane", "policy", "context.must.md");
-      await rm(contextPolicyPath, { force: true });
-      await commitAll(root, "fixture: missing context policy in initialized context workspace");
-
-      io = captureStdIO();
-      try {
-        expect(await runCli(["upgrade", "--yes", "--root", root])).toBe(0);
-      } finally {
-        io.restore();
-      }
-
-      const gatewayText = await readFile(path.join(root, "AGENTS.md"), "utf8");
-      const contextPolicyText = await readFile(contextPolicyPath, "utf8");
-      expect(gatewayText).toContain("@.agentplane/policy/context.must.md");
-      expect(contextPolicyText).toContain("ap context search");
-      expect(contextPolicyText).not.toContain("ap:fragment");
-
-      const execFileAsync = promisify(execFile);
-      const { stdout } = await execFileAsync("node", [".agentplane/policy/check-routing.mjs"], {
-        cwd: root,
-        env: cleanGitEnv(),
-      });
-      expect(String(stdout ?? "")).toContain("policy routing OK");
-    },
-  );
 
   it(
     "upgrade includes runtime .gitignore cache lines in the upgrade commit",
